amazon_toolbox_desktop/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pywebview>=4.4.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
pyinstaller>=6.0.0
